+++ dao-api/src/main/java/com/linkedin/metadata/dao/ingestion/preupdate/RoutingMap.java
@@ -11,5 +11,5 @@ public class RoutingMap {
     SKIP,
     PROCEED
   }
-  public PreUpdateService<? extends Message> preUpdateService;
+  public PreUpdateClient<? extends Message> preUpdateClient;
 }
